chore: upgrade metamask-extension-provider
