+++ package.json
@@ -59,7 +59,6 @@
     "@material-ui/core": "5.0.0-alpha.5",
     "@material-ui/icons": "5.0.0-alpha.4",
     "@material-ui/lab": "5.0.0-alpha.5",
-    "@metamask/inpage-provider": "^7.0.0",
     "@msgpack/msgpack": "^1.12.2",
     "@openzeppelin/contracts": "^3.2.0",
     "@servie/events": "^1.0.0",
@@ -95,12 +94,10 @@
     "clipboard-polyfill": "^3.0.1",
     "d3": "^5.16.0",
     "date-fns": "^2.16.1",
-    "detect-browser": "^5.2.0",
     "elliptic": "^6.5.3",
     "eth-contract-metadata": "^1.15.0",
     "ethereum-blockies": "^0.1.1",
     "event-iterator": "^2.0.0",
-    "extension-port-stream": "^1.0.0",
     "fuse.js": "^6.4.3",
     "gun": "0.2020.520",
     "husky": "^4.3.0",
@@ -115,6 +112,7 @@
     "lodash-es": "npm:lodash@^4.17.11",
     "loglevel": "^1.7.0",
     "matrix-js-sdk": "^9.0.1",
+    "metamask-extension-provider": "^2.0.0",
     "notistack": "^0.9.17",
     "popper.js": "^1.16.1",
     "promievent": "^0.1.3",
@@ -133,7 +131,6 @@
     "react-use": "^15.3.4",
     "react-virtualized-auto-sizer": "^1.0.2",
     "react-window": "^1.8.6",
-    "stream": "^0.0.2",
     "string.prototype.matchall": "^4.0.2",
     "swr": "0.3.8",
     "tiny-secp256k1": "^1.1.5",
